@@ -16,10 +16,10 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 # In Windows, build-name is used as the major, minor, and patch parts
 # of the product and file versions while build-number is used as the build suffix.
-version: 0.4.0-beta7+21
+version: 0.4.0+22
 
 environment:
-  sdk: ^3.8.1
+  sdk: ^3.9.0
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -38,7 +38,7 @@ dependencies:
   audioplayers: ^6.0.0
   gap: ^3.0.1
   dart_random_choice: ^0.2.0
-  flex_color_scheme: ^8.1.0
+  flex_color_scheme: ^8.3.0
   intl: ^0.20.2
   auto_size_text: ^3.0.0
   url_launcher: ^6.1.8
